corrected hw5
